pyproject.toml
[project]
name = "ontology-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.10.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "loguru>=0.7.3",
    "neo4j>=5.28.2",
    "openai>=1.106.1",
    "opensearch-py>=3.0.0",
    "pydantic-settings>=2.10.1",
    "pyvis>=0.3.2",
    "streamlit>=1.49.1",
    "typer>=0.17.4",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
